Backend/batch_layer/requirements.txt
pyspark==2.4.5
pandas==1.0.3
pymongo==3.9.0
azure-storage-blob==12.3.2
numpy==1.18.1